@@ -11,18 +11,23 @@
  * cutoff is not feasible (the data become too large), so we only store it for a single cutoff during surface generation. However, calculating accessibility
  * for additional grids should only take milliseconds.
  *
+ * @param {Number} cutoff
  * @param {Object} surface
  * @param {Object} grid
  * @returns {Number} accessibility
  */
 
-export default function accessibilityForGrid ({cutoff = 60, surface, grid}) {
-  let query = surface.query
-  let accessibility = 0
+export default function accessibilityForGrid ({
+  cutoff = 60,
+  grid,
+  surface
+}) {
+  const query = surface.query
 
+  let accessibility = 0
   for (let y = 0, pixel = 0; y < query.height; y++) {
     for (let x = 0; x < query.width; x++, pixel++) {
-      let travelTime = surface.surface[pixel]
+      const travelTime = surface.surface[pixel]
 
       // ignore unreached locations
       // TODO in OTP/R5 we have a sigmoidal cutoff here to avoid "echoes" of high density locations
@@ -30,15 +35,13 @@ export default function accessibilityForGrid ({cutoff = 60, surface, grid}) {
       // But maybe also we just want to not represent these things as hard edges since accessibility
       // is a continuous phenomenon. No one is saying "ah, rats, it takes 60 minutes and 10 seconds
       // to get work, I have to find a job that's 20 meters closer to home..."
-      if (travelTime > cutoff) continue
-
-      let gridx = x + query.west - grid.west
-      let gridy = y + query.north - grid.north
-
-      // if condition below fails we're off the grid, value is zero, don't bother with calculations
-      if (gridx >= 0 && gridx < grid.width && gridy >= 0 && gridy < grid.height) {
-        // get value of this pixel from grid
-        accessibility += grid.data[gridy * grid.width + gridx]
+      if (travelTime <= cutoff) {
+        const gridx = x + query.west - grid.west
+        const gridy = y + query.north - grid.north
+        if (grid.contains(gridx, gridy)) {
+          // get value of this pixel from grid
+          accessibility += grid.data[gridy * grid.width + gridx]
+        }
       }
     }
   }
