@@ -26,22 +26,20 @@ export default function getSpectrogramData ({origin, query, stopTreeCache, grid}
       x,
       y
     }) => {
-      let gridx = x + query.west - grid.west
-      let gridy = y + query.north - grid.north
+      const gridx = x + query.west - grid.west
+      const gridy = y + query.north - grid.north
 
       // off the grid, return
-      if (gridx < 0 || gridy < 0 || gridx >= grid.width || gridy >= grid.height) return
+      if (grid.contains(gridx, gridy)) {
+        const val = grid.data[gridy * grid.width + gridx]
 
-      const val = grid.data[gridy * grid.width + gridx]
-
-      for (let i = 0; i < travelTimesForDest.length; i++) {
-        const time = travelTimesForDest[i]
-        if (time === 255) continue // unreachable
-
-        if (time < MAX_TRIP_LENGTH) {
-          // time - 1 so areas reachable in 1 minute will be included in output[i][0]
-          // TODO audit all the places we're flooring things
-          output[i][time - 1] += val
+        for (let i = 0; i < travelTimesForDest.length; i++) {
+          const time = travelTimesForDest[i]
+          if (time !== 255 && time < MAX_TRIP_LENGTH) {
+            // time - 1 so areas reachable in 1 minute will be included in output[i][0]
+            // TODO audit all the places we're flooring things
+            output[i][time - 1] += val
+          }
         }
       }
     }})
