@@ -120,7 +120,7 @@ export function computeMedianPixelValue (travelTimes) {
     // odd number, find the middle, keeping in mind the fencepost problem
     return travelTimes[Math.floor(travelTimes.length / 2)]
   } else {
-    let pos = travelTimes.length / 2
+    const pos = travelTimes.length / 2
     // -1 because off-by-one
     return (travelTimes[pos] + travelTimes[pos - 1]) / 2
   }
