@@ -60,7 +60,7 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
     else return p1[1] - p2[1] // then by path index
   })
 
-  let npaths = paths.length
+  const npaths = paths.length
 
   // get (up to) five most common paths
   paths = paths.reduce((out, path, index) => {
@@ -88,13 +88,13 @@ export default function getTransitiveData ({ origin, query, stopTreeCache, netwo
   paths = filterPaths(paths, network)
 
   for (let pidx = 0; pidx < paths.length; pidx++) {
-    let journey = {
+    const journey = {
       journey_id: pidx,
       journey_name: pidx,
       segments: []
     }
 
-    let path = paths[pidx]
+    const path = paths[pidx]
 
     // bail if the path-finding algorithm gave up, this is a bug but we'd rather log a javascript error than crash the tab
     if (path === null) continue
@@ -327,13 +327,13 @@ export function filterPaths (paths, network) {
 
   // eliminate longer paths if there is a shorter path that is a subset (eliminate the short access/egress/transfer leg problem)
   paths = paths.filter((path, i, a) => {
-    for (let otherPath of paths) {
+    for (const otherPath of paths) {
       if (otherPath.length >= path.length) continue // longer paths cannot be subsets. Also don't evaluate the same path.
 
       let otherPathIsSubset = true
-      let routes = path.map(seg => r(seg[1]))
+      const routes = path.map(seg => r(seg[1]))
 
-      for (let seg of otherPath) {
+      for (const seg of otherPath) {
         if (routes.indexOf(r(seg[1])) === -1) {
           otherPathIsSubset = false
           break
@@ -341,7 +341,7 @@ export function filterPaths (paths, network) {
       }
 
       if (otherPathIsSubset) {
-        console.log(`path ${otherPath.map(seg => r(seg[1]))} is subset of path ${routes}`)
+        debug(`path ${otherPath.map(seg => r(seg[1]))} is subset of path ${routes}`)
         return false
       }
     }
@@ -349,7 +349,7 @@ export function filterPaths (paths, network) {
     return true
   })
 
-  console.log(`filtering reduced ${inPathCount} paths to ${pathCountAfterMultiTransfer} after multiple-stop elimination, to ${paths.length} after stemming`)
+  debug(`filtering reduced ${inPathCount} paths to ${pathCountAfterMultiTransfer} after multiple-stop elimination, to ${paths.length} after stemming`)
 
   return paths
 }
@@ -364,8 +364,8 @@ export function bundle (data) {
   // perform hierarchical clustering on journeys
   // see e.g. James et al., _An Introduction to Statistical Learning, with Applications in R_. New York: Springer, 2013, pg. 395.
   // convert to arrays
-  let clusters = data.journeys.map(j => [j])
-  let inputSize = data.journeys.length
+  const clusters = data.journeys.map((j) => [j])
+  const inputSize = data.journeys.length
 
   // prevent infinite loop, makes sense only to loop until there's just one cluster left
   while (clusters.length > 1) {
@@ -376,7 +376,7 @@ export function bundle (data) {
 
     for (let i = 1; i < clusters.length; i++) {
       for (let j = 0; j < i; j++) {
-        let d = clusterDissimilarity(clusters[i], clusters[j], data)
+        const d = clusterDissimilarity(clusters[i], clusters[j], data)
 
         if (d < minDis) {
           minDis = d
@@ -389,11 +389,11 @@ export function bundle (data) {
     debug(`dissimilarity ${minDis}`)
 
     // should we make a cluster?
-    if (minDis > MAXIMUM_DISSIMILARITY) break
-
-    // cluster the least dissimilar clusters
-    clusters[minI] = clusters[minI].concat(clusters[minJ])
-    clusters.splice(minJ, 1) // remove clusters[j]
+    if (minDis <= MAXIMUM_DISSIMILARITY) {
+      // cluster the least dissimilar clusters
+      clusters[minI] = clusters[minI].concat(clusters[minJ])
+      clusters.splice(minJ, 1) // remove clusters[j]
+    }
   }
 
   // merge journeys together
@@ -449,27 +449,27 @@ export function bundle (data) {
 function clusterDissimilarity (c1, c2, data) {
   let ret = 0
 
-  for (let j1 of c1) {
-    for (let j2 of c2) {
+  for (const j1 of c1) {
+    for (const j2 of c2) {
       // if they are not the same length, don't cluster them
       if (j1.segments.length !== j2.segments.length) return Infinity
 
       // otherwise compute maximum dissimilarity of stops at either start or end
       for (let segment = 0; segment < j1.segments.length; segment++) {
-        let s1 = j1.segments[segment]
-        let s2 = j2.segments[segment]
+        const s1 = j1.segments[segment]
+        const s2 = j2.segments[segment]
 
         // if one has a walk segment where the other has a transit segment, that's a problem
         // and these are not comparable
         if (s1.type !== s2.type) return Infinity
 
         // Don't look at walk segments, we only cluster based on the stop positions which we get from transit segments
-        if (s1.type === 'WALK') continue
+        if (s1.type !== 'WALK') {
+          ret = Math.max(ret, segmentDissimilarity(s1, s2, data))
 
-        ret = Math.max(ret, segmentDissimilarity(s1, s2, data))
-
-        // no point in continuing, these won't be merged
-        if (ret > MAXIMUM_DISSIMILARITY) return Infinity
+          // no point in continuing, these won't be merged
+          if (ret > MAXIMUM_DISSIMILARITY) return Infinity
+        }
       }
     }
   }
@@ -479,25 +479,22 @@ function clusterDissimilarity (c1, c2, data) {
 
 /** return the dissimilarity between two individual transit segments (each with only a single pattern, not yet merged) */
 function segmentDissimilarity (s1, s2, data) {
-  let pat1 = data.patterns.find(p => p.pattern_id === s1.pattern_id)
-  let pat2 = data.patterns.find(p => p.pattern_id === s2.pattern_id)
+  const pat1 = data.patterns.find((p) => p.pattern_id === s1.pattern_id)
+  const pat2 = data.patterns.find((p) => p.pattern_id === s2.pattern_id)
 
-  let from1 = pat1.stops[s1.from_stop_index].stop_id
-  let to1 = pat1.stops[s1.to_stop_index].stop_id
-  let from2 = pat2.stops[s2.from_stop_index].stop_id
-  let to2 = pat2.stops[s2.to_stop_index].stop_id
+  const from1 = pat1.stops[s1.from_stop_index].stop_id
+  const to1 = pat1.stops[s1.to_stop_index].stop_id
+  const from2 = pat2.stops[s2.from_stop_index].stop_id
+  const to2 = pat2.stops[s2.to_stop_index].stop_id
 
-  let d1 = stopDistance(from1, from2, data)
-  let d2 = stopDistance(to1, to2, data)
+  const findStop = (id) => data.stops.find((stop) => stop.stop_id === (id + ''))
+  const d1 = stopDistance(findStop(from1), findStop(from2))
+  const d2 = stopDistance(findStop(to1), findStop(to2))
   return Math.max(d1, d2)
 }
 
 /** return the Ersatz (squared) distance between two stops, in undefined units */
-function stopDistance (stop1, stop2, data) {
-  // TODO perhaps this should be optimized?
-  let s1 = data.stops.find(s => s.stop_id === stop1 + '')
-  let s2 = data.stops.find(s => s.stop_id === stop2 + '')
-
-  let cosLat = Math.cos(s1.stop_lat * Math.PI / 180)
+function stopDistance (s1, s2) {
+  const cosLat = Math.cos(s1.stop_lat * Math.PI / 180)
   return Math.pow(s1.stop_lat - s2.stop_lat, 2) + Math.pow(s1.stop_lon * cosLat - s2.stop_lon * cosLat, 2)
 }
