@@ -14,15 +14,22 @@ export function create (data) {
     if (prev < min) min = prev
     if (prev > max) max = prev
   }
-    // parse header
+
+  const width = header[3]
+  const height = header[4]
+
+  // parse header
   return {
     zoom: header[0],
     west: header[1],
     north: header[2],
-    width: header[3],
-    height: header[4],
+    width,
+    height,
     data: array,
     min,
-    max
+    max,
+    contains (x, y) {
+      return x >= 0 && x < width && y >= 0 && y < height
+    }
   }
 }
