@@ -1,11 +1,15 @@
 import WebWorkerPromiseInterface from 'web-worker-promise-interface'
-import {Map as LeafletMap} from 'leaflet' // NB raw leaflet map not react-leaflet
-import {create as createGridFunc} from './grid'
 
+import {create as createGridFunc} from './grid'
+import {latToPixel, lonToPixel} from './mercator'
 import workerHandlers from './worker-handlers'
 
 const imageHeight = 256
 const imageWidth = 256
+
+function assert (e, msg) {
+  if (!e) throw new Error(msg)
+}
 
 export default class Browsochrones extends WebWorkerPromiseInterface {
   originLoaded = false
@@ -30,13 +34,23 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
    * before calling this function if you expect to use it elsewhere
    */
   putGrid (key, grid) {
-    this.work({
+    return this.work({
       command: 'putGrid',
       message: { key, grid }
     })
   }
 
-  drawTile (canvas, {x, y}, zoom) {
+  createTile = (coords, done) => {
+    const tile = document.createElement('canvas')
+    tile.height = imageHeight
+    tile.width = imageWidth
+    this.drawTile(tile, coords, coords.z)
+      .then(() => done(null, tile))
+      .catch((err) => done(err, tile))
+    return tile
+  }
+
+  drawTile = (canvas, {x, y}, zoom) => {
     const scaleFactor = Math.pow(2, zoom - this.query.zoom)
 
     // find top-left coords at zoom 10
@@ -48,7 +62,7 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     const ctx = canvas.getContext('2d')
     const imageData = ctx.createImageData(imageHeight, imageWidth)
 
-    this.work({
+    return this.work({
       command: 'drawTile',
       message: {
         imageData,
@@ -56,15 +70,14 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
         xoffset,
         yoffset
       }
-    }).then(message => {
+    }).then((message) => {
       ctx.putImageData(message, 0, 0)
+      return message
     })
   }
 
-  getAccessibilityForGrid (grid, cutoff = 60) {
-    if (!this.isLoaded()) {
-      return Promise.reject(new Error('Accessibility cannot be computed before generating a surface.'))
-    }
+  async getAccessibilityForGrid (grid, cutoff = 60) {
+    assert(this.isLoaded(), 'Accessibility cannot be computed before generating a surface.')
 
     return this.work({
       command: 'accessibilityForGrid',
@@ -74,10 +87,8 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     })
   }
 
-  getSpectrogramData (grid) {
-    if (!this.isLoaded()) {
-      return Promise.reject(new Error('Accessibility cannot be computed before generating a surface.'))
-    }
+  async getSpectrogramData (grid) {
+    assert(this.isLoaded(), 'Accessibility cannot be computed before generating a surface.')
 
     return this.work({
       command: 'getSpectrogramData',
@@ -87,10 +98,8 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     })
   }
 
-  generateDestinationData ({ from, to }) {
-    if (!this.isLoaded()) {
-      return Promise.reject(new Error('Transitive data cannot be generated if Browsochrones is not fully loaded.'))
-    }
+  async generateDestinationData ({ from, to }) {
+    assert(this.isLoaded(), 'Transitive data cannot be generated if Browsochrones is not fully loaded.')
 
     return this.work({
       command: 'generateDestinationData',
@@ -119,38 +128,30 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
     })
   }
 
-  generateSurface (which = 'MEDIAN') {
-    if (!this.isLoaded()) {
-      return Promise.reject(new Error('Surface cannot be generated if Browsochrones is not fully loaded.'))
-    }
+  async generateSurface (which = 'MEDIAN') {
+    assert(this.isLoaded(), 'Surface cannot be generated if Browsochrones is not fully loaded.')
 
     return this.work({
       command: 'generateSurface',
       message: {
         which
       }
-    }).then(message => {
+    }).then((message) => {
       this.surfaceLoaded = true
     })
   }
 
   latLonToOriginPoint ({lat, lon, lng}) {
-    lon = lon || lng
-
-    if (!this.query) throw new Error('Cannot convert lat/lon without query.')
-
-    let {north, west, zoom} = this.query
-
-    let {x, y} = LeafletMap.prototype.project([lat, lon], zoom)
-
-    let ret = {
+    assert(this.query, 'Cannot convert lat/lon without query.')
+    const {north, west, zoom} = this.query
+    const x = latToPixel(lat, zoom)
+    const y = lonToPixel(lon || lng, zoom)
+    const ret = {
       // TODO should these be rounded instead of floored?
       x: (x - west) | 0,
       y: (y - north) | 0
     }
-
-    if (!this.pointInQueryBounds(ret)) throw new Error('Point out of query bounds!')
-
+    assert(this.pointInQueryBounds(ret), 'Point out of query bounds!')
     return ret
   }
 
@@ -161,9 +162,7 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
    * @returns {Object} point
    */
   pixelToOriginPoint ({x, y}, currentZoom) {
-    if (!this.query) {
-      throw new Error('Cannot convert point without query.')
-    }
+    assert(this.query, 'Cannot convert point without query.')
 
     const {north, west, zoom} = this.query
     const scale = Math.pow(2, zoom - currentZoom)
@@ -218,10 +217,8 @@ export default class Browsochrones extends WebWorkerPromiseInterface {
    *
    * @param {ArrayBuffer} data
    */
-  setStopTrees (arrayBuffer) {
-    if (!this.query) {
-      return Promise.reject(new Error('Query must be loaded before generating the stop tree cache.'))
-    }
+  async setStopTrees (arrayBuffer) {
+    assert(this.query, 'Query must be loaded before generating the stop tree cache.')
 
     return this.work({
       command: 'setStopTreeCache',
