@@ -6,13 +6,13 @@
 import Color from 'color'
 import slice from 'lodash.slice'
 
-let colorTable = new Uint8Array(255 * 4)
+const colorTable = new Uint8Array(255 * 4)
 
 for (let i = 0; i < 254; i++) {
-  let val = Math.min(i, 120)
+  const val = Math.min(i, 120)
 
   // 15 minutes of travel time represents 45 degrees on the color wheel
-  let color = Color().hsl(0, 1, 70)
+  const color = Color().hsl(0, 1, 70)
   color.rotate(45 * (val / 15 | 0))
 
   // saturation represents gradation within a 15-minute period
@@ -32,8 +32,8 @@ export default function isochroneTile (imageData, {height, scaleFactor, surface,
   for (let yp = 0, pixel = 0; yp < imageData.height; yp++) {
     for (let xp = 0; xp < imageData.width; xp++, pixel = (pixel + 1) | 0) {
       // figure out where xp and yp fall on the surface
-      let xpsurf = (xp / scaleFactor + xoffset) | 0
-      let ypsurf = (yp / scaleFactor + yoffset) | 0
+      const xpsurf = (xp / scaleFactor + xoffset) | 0
+      const ypsurf = (yp / scaleFactor + yoffset) | 0
 
       let val
       if (xpsurf < 0 || xpsurf > width || ypsurf < 0 || ypsurf > height) {
@@ -42,11 +42,11 @@ export default function isochroneTile (imageData, {height, scaleFactor, surface,
         val = surface[ypsurf * width + xpsurf]
       }
 
-      let colorScheme = colorScheme || function (val) {
+      const _colorScheme = colorScheme || function (val) {
         return slice(colorTable, val * 4, val * 4 + 4)
       }
 
-      let col = colorScheme(val)
+      const col = _colorScheme(val)
       // 50% transparent yellow (#ddddaa)
       imageData.data[pixel * 4] = col[0]
       imageData.data[pixel * 4 + 1] = col[1]
