@@ -9,9 +9,9 @@ export function lonToPixel (lon, zoom) {
 
 /** convert latitude to pixel value */
 export function latToPixel (lat, zoom) {
-  let invCos = 1 / Math.cos(radians(lat))
-  let tan = Math.tan(radians(lat))
-  let ln = Math.log(tan + invCos)
+  const invCos = 1 / Math.cos(radians(lat))
+  const tan = Math.tan(radians(lat))
+  const ln = Math.log(tan + invCos)
   // return int
   return ((1 - ln / Math.PI) * Math.pow(2, zoom - 1) * 256) | 0
 }
@@ -23,7 +23,7 @@ export function pixelToLon (x, zoom) {
 
 /** convert pixel value to latitude */
 export function pixelToLat (y, zoom) {
-  let tile = y / 256
+  const tile = y / 256
   return degrees(Math.atan(Math.sinh(Math.PI - tile * Math.PI * 2 / Math.pow(2, zoom))))
 }
 
