@@ -23,7 +23,7 @@ export function create (data, point) {
   // de-delta-code non transit timess
   let offset = 0
   origin.radius = data[offset++]
-  let nPixels = Math.pow(origin.radius * 2 + 1, 2)
+  const nPixels = Math.pow(origin.radius * 2 + 1, 2)
 
   for (let pixel = 0, nonTransitTime = 0; pixel < nPixels; pixel++) {
     nonTransitTime += data[offset] // don't increment offset here, will be incremented on write
@@ -31,8 +31,8 @@ export function create (data, point) {
   }
 
   // build index
-  let nStops = data[offset++]
-  let nMinutes = data[offset++]
+  const nStops = data[offset++]
+  const nMinutes = data[offset++]
   origin.nStops = nStops
   origin.nMinutes = nMinutes
 
@@ -60,10 +60,10 @@ export function create (data, point) {
     offset += nMinutes * ITERATION_WIDTH
 
     // read paths
-    let nPaths = data[offset++]
+    const nPaths = data[offset++]
 
     for (let path = 0; path < nPaths; path++) {
-      let pathLen = data[offset++]
+      const pathLen = data[offset++]
       offset += pathLen * 3 // 3 ints for each segment: board, pattern, and alight
     }
   }
