@@ -17,7 +17,7 @@ import fill from 'lodash.fill'
 import {getNonTransitTime, ITERATION_WIDTH} from './origin'
 
 export default function propagate ({ query, stopTreeCache, origin, callback }) {
-  let {nMinutes} = origin
+  const {nMinutes} = origin
   const travelTimesForDest = new Uint8Array(nMinutes) // the total travel time per iteration to reach a particular destination
   const waitTimesForDest = new Uint8Array(nMinutes) // wait time per iteration for particular destination
   const inVehicleTravelTimesForDest = new Uint8Array(nMinutes) // in-vehicle travel time per destination
