@@ -1,4 +1,4 @@
-import jsolines, { getContour } from 'jsolines'
+import jsolines, {getContour} from 'jsolines'
 import {createHandler} from 'web-worker-promise-interface'
 
 import accessibilityForGrid from './accessibility-for-grid'
@@ -57,12 +57,12 @@ module.exports = createHandler({
   generateDestinationData (ctx, message) {
     const { to, from } = message
 
-    let travelTime = ctx.surface.surface[to.y * ctx.query.width + to.x]
-    let waitTime = ctx.surface.waitTimes[to.y * ctx.query.width + to.x]
+    const travelTime = ctx.surface.surface[to.y * ctx.query.width + to.x]
+    const waitTime = ctx.surface.waitTimes[to.y * ctx.query.width + to.x]
     // NB separate walk time surface because some summary statistics don't preserve stat(wait) +
     //   stat(walk) + stat(inVehicle) = stat(total)
-    let walkTime = ctx.surface.walkTimes[to.y * ctx.query.width + to.x]
-    let inVehicleTravelTime = ctx.surface.inVehicleTravelTimes[to.y * ctx.query.width + to.x]
+    const walkTime = ctx.surface.walkTimes[to.y * ctx.query.width + to.x]
+    const inVehicleTravelTime = ctx.surface.inVehicleTravelTimes[to.y * ctx.query.width + to.x]
 
     return {
       paths: getPaths({
